Updates to NetworkClient so that it doesn't have to do as much work
